packages/tf_trade/requirements.txt
telethon
python-dotenv
colorlog
docker
torch
scipy
stable-baselines3[extra]
rpyc
git+https://github.com/fortesenselabs/trade_flow.git
joblib
aiodbm
sqlalchemy

# dev
ta
seaborn